rename: types 파일명 수정 + 추가 이미지 다운
