+++ src/type/StatusBarType.types.ts
@@ -1,0 +1,3 @@
+export type StatusBarProps = {
+  theme?: 'brown' | 'white';
+};
